(BSR)[PRO] refacto: remove phone number input
