--- pro/src/ui-kit/form/PhoneNumberInput/CodeCountrySelect/CountryCodeSelect.module.scss
@@ -1,91 +1,0 @@
-@use "styles/mixins/_forms.scss" as formsM;
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/variables/_forms.scss" as forms;
-@use "styles/mixins/_size.scss" as size;
-
-$icon-width: rem.torem(16px);
-$icon-padding-left: rem.torem(16px);
-$space-icon-text: rem.torem(8px);
-$focus-border-width: rem.torem(2px);
-$base-border-width: rem.torem(1px);
-$input-height: calc(
-  #{size.$input-min-height} - (2 * #{size.$input-border-width})
-);
-$input-width: rem.torem(86px);
-$text-padding-left: calc(
-  #{$icon-width} + #{$icon-padding-left} + #{$space-icon-text}
-);
-
-.select-input-wrapper {
-  position: relative;
-  border-radius: rem.torem(forms.$input-border-radius) 0 0
-    rem.torem(forms.$input-border-radius);
-  margin: 0 $base-border-width;
-
-  &:focus,
-  &:focus-within {
-    border: solid var(--color-border-focused) $base-border-width;
-    border-radius: rem.torem(forms.$input-border-radius) 0 0
-      rem.torem(forms.$input-border-radius);
-    background-color: transparent;
-    box-shadow: inset 0 0 0 rem.torem(1px) var(--color-border-focused);
-
-    .select-input {
-      height: calc(#{$input-height} - #{$focus-border-width});
-      min-height: calc(#{$input-height} - #{$focus-border-width});
-      padding-left: calc(#{$text-padding-left} - (#{$focus-border-width} / 2));
-    }
-
-    .select-input-icon {
-      left: calc(#{$icon-padding-left} - (#{$focus-border-width} / 2));
-    }
-  }
-
-  &:hover {
-    border-radius: rem.torem(forms.$input-border-radius) 0 0
-      rem.torem(forms.$input-border-radius);
-  }
-}
-
-.select-input {
-  @include formsM.input-theme;
-  @include fonts.body-accent;
-
-  padding-left: $text-padding-left;
-  padding-right: rem.torem(8px);
-  width: $input-width;
-  height: $input-height;
-  min-height: $input-height;
-  margin: 1px;
-  background-color: transparent;
-
-  // -- override default select style --
-  border: unset;
-  box-shadow: unset;
-
-  @include formsM.input-theme-nested;
-
-  &:hover,
-  &:hover:not(:focus),
-  &:focus {
-    box-shadow: none;
-    outline: none;
-    border: none;
-  }
-
-  &-icon {
-    position: absolute;
-    top: 0;
-    left: $icon-padding-left;
-    height: 100%;
-    width: $icon-width;
-    fill: var(--color-icon-default);
-    transform: rotate(90deg);
-    pointer-events: none;
-
-    &-disabled {
-      fill: var(--color-icon-disabled);
-    }
-  }
-}
