--- pro/src/ui-kit/form/PhoneNumberInput/PhoneNumberInput.module.scss
@@ -1,68 +1,0 @@
-@use "./CodeCountrySelect/CountryCodeSelect.module.scss" as countryCodeSelect;
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/variables/_forms.scss" as forms;
-@use "styles/mixins/_forms.scss" as formsM;
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_a11y.scss" as a11y;
-
-$input-phone-number-padding-left: calc(
-  countryCodeSelect.$input-width + rem.torem(8px) +
-    countryCodeSelect.$focus-border-width
-);
-
-.phone-number {
-  width: 100%;
-  margin-bottom: rem.torem(8px);
-
-  &-input {
-    padding-left: calc($input-phone-number-padding-left + rem.torem(1px));
-
-    &:focus {
-      padding-left: $input-phone-number-padding-left;
-    }
-
-    &-wrapper {
-      position: relative;
-      margin-bottom: rem.torem(16px);
-    }
-
-    &-legend {
-      display: flex;
-      width: 100%;
-      justify-content: space-between;
-    }
-
-    &-footer {
-      @include formsM.field-layout-footer;
-    }
-
-    &-error {
-      flex: 1;
-
-      svg {
-        flex: 0 0 15px;
-      }
-    }
-  }
-}
-
-.country-code-select {
-  position: absolute;
-  top: 0;
-  left: -1px;
-}
-
-.visually-hidden {
-  @include a11y.visually-hidden;
-}
-
-.phone-format {
-  @include fonts.body-xs;
-
-  color: var(--color-text-subtle);
-  margin-bottom: rem.torem(forms.$label-space-before-input);
-}
-
-.phone-number-inpus {
-  position: relative;
-}
