--- pro/src/ui-kit/form/PhoneNumberInput/types.ts
@@ -1,6 +1,0 @@
-import type { CountryCode } from 'libphonenumber-js'
-
-export type PhoneCodeSelectOption = {
-  label: string
-  value: CountryCode
-}
